Begin update of Apollo Server
- `context` function has been moved to the express app middleware, not
  the ApolloServer instance
- Refactored error functions to use redefined GraphQL Error templates,
  as `ApolloError` was removed
- Removed `gql` constant, as Apollo Server no longer uses it

Testing is still necessary to ensure things are working as intended.
