+++ package.json
@@ -11,9 +11,10 @@
 		"build": "npm run migrate"
 	},
 	"dependencies": {
-		"apollo-server-core": "^3.5.0",
-		"apollo-server-express": "^3.5.0",
+		"@apollo/server": "^4.3.2",
+		"body-parser": "^1.20.1",
 		"connect-session-sequelize": "^7.1.5",
+		"cors": "^2.8.5",
 		"dotenv": "^16.0.3",
 		"esm": "^3.2.25",
 		"express": "^4.18.2",
