+++ src/graphql/resolvers/Mutation/loginWithGoogle.js
@@ -1,4 +1,4 @@
-import { AuthenticationError } from "apollo-server-core";
+import { AuthenticationError } from "../../index";
 
 import validateGoogleIDToken from "../../../utils/validateGoogleIDToken";
